functions/orphan-cleanup/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-storage==2.*
